+++ Config/module.xml
@@ -7,7 +7,7 @@
     <descriptive locale="en_US">
         <title>Tabs</title>
     </descriptive>
-    <version>2.4.3</version>
+    <version>2.4.2</version>
     <author>
         <name>Michaël Espeche</name>
         <email>mespeche@openstudio.fr</email>
