+++ Controller/Base/BaseTabsController.php
@@ -21,43 +21,33 @@
 /*                                                                                   */
 /*************************************************************************************/
 
-namespace Tabs\Controller;
+namespace Tabs\Controller\Base;
 
-use Propel\Runtime\Exception\PropelException;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Tabs\Event\TabsDeleteEvent;
 use Tabs\Event\TabsEvent;
-use Tabs\Form\TabsCategoryForm;
-use Tabs\Form\TabsContentForm;
-use Tabs\Form\TabsFolderForm;
-use Tabs\Form\TabsProductForm;
-use Tabs\Model\CategoryAssociatedTabQuery;
-use Tabs\Model\ContentAssociatedTabQuery;
-use Tabs\Model\FolderAssociatedTabQuery;
 use Tabs\Model\ProductAssociatedTab;
 use Tabs\Model\ProductAssociatedTabQuery;
 use Thelia\Controller\Admin\AbstractCrudController;
 use Thelia\Core\Event\UpdatePositionEvent;
 use Thelia\Core\HttpFoundation\Response;
-use Thelia\Core\Security\AccessManager;
 use Thelia\Form\CategoryModificationForm;
 use Thelia\Form\ContentModificationForm;
-use Thelia\Form\Exception\FormValidationException;
 use Thelia\Form\FolderModificationForm;
 use Thelia\Form\ProductModificationForm;
-use Thelia\Model\ProductQuery;
 use Thelia\Model\CategoryQuery;
 use Thelia\Model\ContentQuery;
 use Thelia\Model\FolderQuery;
+use Thelia\Model\ProductQuery;
 use Thelia\Tools\URL;
 
 /**
- * Class TabsController
- * @package Tabs\Controller\Admin
+ * Class BaseTabsController
+ * @package Tabs\Controller\Base
  * @author Michaël Espeche <mespeche@openstudio.fr>
  */
-class TabsController extends AbstractCrudController
+class BaseTabsController extends AbstractCrudController
 {
 	static $possibleVars = [ 'product_id', 'category_id', 'content_id', 'folder_id'];
 
@@ -107,433 +97,11 @@ class TabsController extends AbstractCrudController
 		return $this->render('tabs-config');
 	}
 
-	public function manageTabsContentAssociation($contentId)
-	{
-
-		if (null !== $response = $this->checkAuth(array(), array('Tabs'), AccessManager::UPDATE)) {
-			return $response;
-		}
-
-		$tabId = $this->getRequest()->get('tab_id', null);
-		if (null === $tabId) {
-			return $this->createNewTabContentAssociation($contentId);
-		} else {
-			return $this->updateTabContentAssociation($tabId);
-		}
-
-	}
-
-	public function manageTabsCategoryAssociation($categoryId)
-	{
-
-		if (null !== $response = $this->checkAuth(array(), array('Tabs'), AccessManager::UPDATE)) {
-			return $response;
-		}
-
-		$tabId = $this->getRequest()->get('tab_id', null);
-		if (null === $tabId) {
-			return $this->createNewTabCategoryAssociation($categoryId);
-		} else {
-			return $this->updateTabCategoryAssociation($tabId);
-		}
-
-	}
-
-	public function manageTabsFolderAssociation($folderId)
-	{
-
-		if (null !== $response = $this->checkAuth(array(), array('Tabs'), AccessManager::UPDATE)) {
-			return $response;
-		}
-
-		$tabId = $this->getRequest()->get('tab_id', null);
-		if (null === $tabId) {
-			return $this->createNewTabFolderAssociation($folderId);
-		} else {
-			return $this->updateTabFolderAssociation($tabId);
-		}
-
-	}
-
-	public function createNewTabContentAssociation($contentId)
-	{
-
-		$tabsContentForm = new TabsContentForm($this->getRequest());
-
-		$message = false;
-
-		try {
-			$content = ContentQuery::create()->findPk($contentId);
-
-			if (null === $content) {
-				throw new \InvalidArgumentException(sprintf("%d content id does not exist", $contentId));
-			}
-
-			$form = $this->validateForm($tabsContentForm);
-
-			$event = $this->createEventInstance($form->getData());
-			$event->setContentId($content->getId());
-
-			$this->dispatch(TabsEvent::TABS_CONTENT_CREATE, $event);
-
-			return $this->generateSuccessRedirect($tabsContentForm);
-
-		} catch (FormValidationException $e) {
-			$message = sprintf("Please check your input: %s", $e->getMessage());
-		} catch (PropelException $e) {
-			$message = $e->getMessage();
-		} catch (\Exception $e) {
-			$message = sprintf("Sorry, an error occured: %s", $e->getMessage() . " " . $e->getFile());
-		}
-
-		if ($message !== false) {
-			\Thelia\Log\Tlog::getInstance()->error(sprintf("Error during tabs content association process : %s.", $message));
-
-			$tabsContentForm->setErrorMessage($message);
-
-			$this->getParserContext()
-				->addForm($tabsContentForm)
-				->setGeneralError($message);
-		}
-
-		return $this->updateAction();
-	}
-
-	public function updateTabContentAssociation($tabId)
-	{
-
-		$tabsContentForm = new TabsContentForm($this->getRequest());
-
-		$message = false;
-
-		try {
-			$tab = ContentAssociatedTabQuery::create()->findPk($tabId);
-
-			if (null === $tab) {
-				throw new \InvalidArgumentException(sprintf("%d tab id does not exist", $tabId));
-			}
-
-			$form = $this->validateForm($tabsContentForm);
-
-			$event = $this->createEventInstance($form->getData());
-			$event->setTabId($tab->getId());
-			$event->setContentId($tab->getContentId());
-
-			$this->dispatch(TabsEvent::TABS_CONTENT_UPDATE, $event);
-
-			return $this->generateSuccessRedirect($tabsContentForm);
-
-		} catch (FormValidationException $e) {
-			$message = sprintf("Please check your input: %s", $e->getMessage());
-		} catch (PropelException $e) {
-			$message = $e->getMessage();
-		} catch (\Exception $e) {
-			$message = sprintf("Sorry, an error occured: %s", $e->getMessage() . " " . $e->getFile());
-		}
-
-		if ($message !== false) {
-			\Thelia\Log\Tlog::getInstance()->error(sprintf("Error during tabs content association process : %s.", $message));
-
-			$tabsContentForm->setErrorMessage($message);
-
-			$this->getParserContext()
-				->addForm($tabsContentForm)
-				->setGeneralError($message);
-		}
-
-		return $this->updateAction();
-
-	}
-
-	public function manageTabsProductAssociation($productId)
-	{
-
-		if (null !== $response = $this->checkAuth(array(), array('Tabs'), AccessManager::UPDATE)) {
-			return $response;
-		}
-
-		$tabId = $this->getRequest()->get('tab_id', null);
-		if (null === $tabId) {
-			return $this->createNewTabProductAssociation($productId);
-		} else {
-			return $this->updateTabProductAssociation($tabId);
-		}
-
-	}
-
-	public function createNewTabProductAssociation($productId)
-	{
-
-		$tabsProductForm = new TabsProductForm($this->getRequest());
-
-		$message = false;
-
-		try {
-			$product = ProductQuery::create()->findPk($productId);
-
-			if (null === $product) {
-				throw new \InvalidArgumentException(sprintf("%d content id does not exist", $productId));
-			}
-
-			$form = $this->validateForm($tabsProductForm);
-
-			$event = $this->createEventInstance($form->getData());
-			$event->setProductId($product->getId());
-
-			$this->dispatch(TabsEvent::TABS_PRODUCT_CREATE, $event);
-
-			return $this->generateSuccessRedirect($tabsProductForm);
-		} catch (FormValidationException $e) {
-			$message = sprintf("Please check your input: %s", $e->getMessage());
-		} catch (PropelException $e) {
-			$message = $e->getMessage();
-		} catch (\Exception $e) {
-			$message = sprintf("Sorry, an error occured: %s", $e->getMessage() . " " . $e->getFile());
-		}
-
-		if ($message !== false) {
-			\Thelia\Log\Tlog::getInstance()->error(sprintf("Error during tabs product association process : %s.", $message));
-
-			$tabsProductForm->setErrorMessage($message);
-
-			$this->getParserContext()
-				->addForm($tabsProductForm)
-				->setGeneralError($message);
-		}
-
-		return $this->updateAction();
-	}
-
-	public function updateTabProductAssociation($tabId)
-	{
-
-		$tabsProductForm = new TabsProductForm($this->getRequest());
-
-		$message = false;
-
-		try {
-			$tab = ProductAssociatedTabQuery::create()->findPk($tabId);
-
-			if (null === $tab) {
-				throw new \InvalidArgumentException(sprintf("%d tab id does not exist", $tabId));
-			}
-
-			$form = $this->validateForm($tabsProductForm);
-
-			$event = $this->createEventInstance($form->getData());
-			$event->setTabId($tab->getId());
-			$event->setProductId($tab->getProductId());
-
-			$this->dispatch(TabsEvent::TABS_PRODUCT_UPDATE, $event);
-
-			return $this->generateSuccessRedirect($tabsProductForm);
-		} catch (FormValidationException $e) {
-			$message = sprintf("Please check your input: %s", $e->getMessage());
-		} catch (PropelException $e) {
-			$message = $e->getMessage();
-		} catch (\Exception $e) {
-			$message = sprintf("Sorry, an error occured: %s", $e->getMessage() . " " . $e->getFile());
-		}
-
-		if ($message !== false) {
-			\Thelia\Log\Tlog::getInstance()->error(sprintf("Error during tabs product association process : %s.", $message));
-
-			$tabsProductForm->setErrorMessage($message);
-
-			$this->getParserContext()
-				->addForm($tabsProductForm)
-				->setGeneralError($message);
-		}
-
-		return $this->updateAction();
-
-	}
-
-	public function createNewTabCategoryAssociation($categoryId)
-	{
-
-		$tabsCategoryForm = new TabsCategoryForm($this->getRequest());
-
-		$message = false;
-
-		try {
-			$category = CategoryQuery::create()->findPk($categoryId);
-
-			if (null === $category) {
-				throw new \InvalidArgumentException(sprintf("%d category id does not exist", $categoryId));
-			}
-
-			$form = $this->validateForm($tabsCategoryForm);
-
-			$event = $this->createEventInstance($form->getData());
-			$event->setCategoryId($category->getId());
-
-			$this->dispatch(TabsEvent::TABS_CATEGORY_CREATE, $event);
-
-			return $this->generateSuccessRedirect($tabsCategoryForm);
-
-		} catch (FormValidationException $e) {
-			$message = sprintf("Please check your input: %s", $e->getMessage());
-		} catch (PropelException $e) {
-			$message = $e->getMessage();
-		} catch (\Exception $e) {
-			$message = sprintf("Sorry, an error occured: %s", $e->getMessage() . " " . $e->getFile());
-		}
-
-		if ($message !== false) {
-			\Thelia\Log\Tlog::getInstance()->error(sprintf("Error during tabs category association process : %s.", $message));
-
-			$tabsCategoryForm->setErrorMessage($message);
-
-			$this->getParserContext()
-				->addForm($tabsCategoryForm)
-				->setGeneralError($message);
-		}
-
-		return $this->updateAction();
-	}
-
-	public function updateTabCategoryAssociation($tabId)
-	{
-
-		$tabsCategoryForm = new TabsCategoryForm($this->getRequest());
-
-		$message = false;
-
-		try {
-			$tab = CategoryAssociatedTabQuery::create()->findPk($tabId);
-
-			if (null === $tab) {
-				throw new \InvalidArgumentException(sprintf("%d tab id does not exist", $tabId));
-			}
-
-			$form = $this->validateForm($tabsCategoryForm);
-
-			$event = $this->createEventInstance($form->getData());
-			$event->setTabId($tab->getId());
-			$event->setCategoryId($tab->getCategoryId());
-
-			$this->dispatch(TabsEvent::TABS_CATEGORY_UPDATE, $event);
-
-			return $this->generateSuccessRedirect($tabsCategoryForm);
-
-		} catch (FormValidationException $e) {
-			$message = sprintf("Please check your input: %s", $e->getMessage());
-		} catch (PropelException $e) {
-			$message = $e->getMessage();
-		} catch (\Exception $e) {
-			$message = sprintf("Sorry, an error occured: %s", $e->getMessage() . " " . $e->getFile());
-		}
-
-		if ($message !== false) {
-			\Thelia\Log\Tlog::getInstance()->error(sprintf("Error during tabs folder association process : %s.", $message));
-
-			$tabsCategoryForm->setErrorMessage($message);
-
-			$this->getParserContext()
-				->addForm($tabsCategoryForm)
-				->setGeneralError($message);
-		}
-
-		return $this->updateAction();
-
-	}
-
-	public function createNewTabFolderAssociation($folderId)
-	{
-
-		$tabsFolderForm = new TabsFolderForm($this->getRequest());
-
-		$message = false;
-
-		try {
-			$folder = FolderQuery::create()->findPk($folderId);
-
-			if (null === $folder) {
-				throw new \InvalidArgumentException(sprintf("%d folder id does not exist", $folderId));
-			}
-
-			$form = $this->validateForm($tabsFolderForm);
-
-			$event = $this->createEventInstance($form->getData());
-			$event->setFolderId($folder->getId());
-
-			$this->dispatch(TabsEvent::TABS_FOLDER_CREATE, $event);
-
-			return $this->generateSuccessRedirect($tabsFolderForm);
-
-		} catch (FormValidationException $e) {
-			$message = sprintf("Please check your input: %s", $e->getMessage());
-		} catch (PropelException $e) {
-			$message = $e->getMessage();
-		} catch (\Exception $e) {
-			$message = sprintf("Sorry, an error occured: %s", $e->getMessage() . " " . $e->getFile());
-		}
-
-		if ($message !== false) {
-			\Thelia\Log\Tlog::getInstance()->error(sprintf("Error during tabs folder association process : %s.", $message));
-
-			$tabsFolderForm->setErrorMessage($message);
-
-			$this->getParserContext()
-				->addForm($tabsFolderForm)
-				->setGeneralError($message);
-		}
-
-		return $this->updateAction();
-	}
-
-	public function updateTabFolderAssociation($tabId)
-	{
-
-		$tabsFolderForm = new TabsFolderForm($this->getRequest());
-
-		$message = false;
-
-		try {
-			$tab = FolderAssociatedTabQuery::create()->findPk($tabId);
-
-			if (null === $tab) {
-				throw new \InvalidArgumentException(sprintf("%d tab id does not exist", $tabId));
-			}
-
-			$form = $this->validateForm($tabsFolderForm);
-
-			$event = $this->createEventInstance($form->getData());
-			$event->setTabId($tab->getId());
-			$event->setFolderId($tab->getFolderId());
-
-			$this->dispatch(TabsEvent::TABS_FOLDER_UPDATE, $event);
-
-			return $this->generateSuccessRedirect($tabsFolderForm);
-
-		} catch (FormValidationException $e) {
-			$message = sprintf("Please check your input: %s", $e->getMessage());
-		} catch (PropelException $e) {
-			$message = $e->getMessage();
-		} catch (\Exception $e) {
-			$message = sprintf("Sorry, an error occured: %s", $e->getMessage() . " " . $e->getFile());
-		}
-
-		if ($message !== false) {
-			\Thelia\Log\Tlog::getInstance()->error(sprintf("Error during tabs folder association process : %s.", $message));
-
-			$tabsFolderForm->setErrorMessage($message);
-
-			$this->getParserContext()
-				->addForm($tabsFolderForm)
-				->setGeneralError($message);
-		}
-
-		return $this->updateAction();
-
-	}
-
 	/**
 	 * @param $data
 	 * @return \Tabs\Event\TabsEvent
 	 */
-	private function createEventInstance($data)
+	protected function createEventInstance($data)
 	{
 
 		$tabsAssociationEvent = new TabsEvent(
